+++ packages/cldr-core/__tests__/api/numbers/adjust.test.ts
@@ -1,10 +1,22 @@
 import { numbersApi } from '../../_helpers';
-import { DecimalAdjustOptions } from '../../../src';
+import { Decimal, DecimalAdjustOptions } from '../../../src';
 
 const API = numbersApi('en');
 
 const adjust = (s: string, opts?: DecimalAdjustOptions): string =>
   API.adjustDecimal(s, opts).toString();
+
+const parse = (n: number | string): Decimal => API.parseDecimal(n);
+
+test('parse decimal', () => {
+  let d: Decimal;
+
+  d = parse('1.0');
+  expect(d.toString()).toEqual('1.0');
+
+  d = parse(3.14159);
+  expect(d.toString()).toEqual('3.14159');
+});
 
 test('adjust fractions', () => {
   let s: string;